refactor: update the default sign_data (#560)
* refactor: update the default sign_data

* clean keccak_inputs_tx_circuit

* chore: clippy fix

* add sanity check

* fix

---------
