+++ eth-types/src/sign_types.rs
@@ -1,8 +1,13 @@
 //! secp256k1 signature types and helper functions.
 
-use crate::{ToBigEndian, Word};
+use crate::{
+    address,
+    geth_types::{Transaction, TxType},
+    word, ToBigEndian, Word, H256,
+};
 use ethers_core::{
-    types::{Address, Bytes},
+    k256::ecdsa::SigningKey,
+    types::{Address, Bytes, Signature, TransactionRequest, U256},
     utils::keccak256,
 };
 use halo2_proofs::{
@@ -18,7 +23,6 @@ use halo2_proofs::{
 };
 use lazy_static::lazy_static;
 use num_bigint::BigUint;
-use sha3::{Digest, Keccak256};
 use subtle::CtOption;
 
 /// Do a secp256k1 signature with a given randomness value.
@@ -61,6 +65,39 @@ pub struct SignData {
     pub msg_hash: secp256k1::Fq,
 }
 
+/// Generate a dummy pre-eip155 tx in which
+/// (nonce=0, gas=0, gas_price=0, to=0, value=0, data="")
+/// using the dummy private key = 1
+pub fn get_dummy_tx() -> (TransactionRequest, Signature) {
+    let mut sk_be_scalar = [0u8; 32];
+    sk_be_scalar[31] = 1_u8;
+
+    let sk = SigningKey::from_bytes(&sk_be_scalar).expect("sign key = 1");
+    let wallet = ethers_signers::Wallet::from(sk);
+
+    let tx = TransactionRequest::new()
+        .nonce(0)
+        .gas(0)
+        .gas_price(U256::zero())
+        .to(Address::zero())
+        .value(U256::zero())
+        .data(Bytes::default());
+    let sighash: H256 = keccak256(tx.rlp_unsigned()).into();
+
+    let sig = wallet.sign_hash(sighash);
+    assert_eq!(sig.v, 28);
+    assert_eq!(
+        sig.r,
+        word!("4faabf49beea23083894651a6f34baaf3dc29b396fb5baf8b8454773f328df61")
+    );
+    assert_eq!(
+        sig.s,
+        word!("0x75ae2dd5e4e688c9dbc6db7e75bafcb04ea141ca20332be9809a444d541272c1")
+    );
+
+    (tx, sig)
+}
+
 impl SignData {
     /// Recover address of the signature
     pub fn get_addr(&self) -> Address {
@@ -74,37 +111,24 @@ impl SignData {
 }
 
 lazy_static! {
-    // FIXME: use Transaction::dummy().sign_data() instead when we merged the develop branch
+    /// This is the sign data of default padding tx
     static ref SIGN_DATA_DEFAULT: SignData = {
-        let generator = Secp256k1Affine::generator();
-        let sk = secp256k1::Fq::one();
-        let pk = generator * sk;
-        let pk = pk.to_affine();
-        let msg = b"1";
-        // let msg = TransactionRequest::new()
-        //     .nonce(0)
-        //     .gas(0)
-        //     .gas_price(U256::zero())
-        //     .to(Address::zero())
-        //     .value(U256::zero())
-        //     .data(Bytes::default())
-        //     .chain_id(1)
-        //     .rlp().to_vec();
-        let msg_hash: [u8; 32] = Keccak256::digest(msg)
-            .as_slice()
-            .to_vec()
-            .try_into()
-            .expect("hash length isn't 32 bytes");
-        let msg_hash = secp256k1::Fq::from_bytes(&msg_hash).unwrap();
-        let randomness = secp256k1::Fq::one();
-        let (sig_r, sig_s, v) = sign(randomness, sk, msg_hash);
+        let (tx_req, sig) = get_dummy_tx();
+        let tx = Transaction {
+            tx_type: TxType::PreEip155,
+            rlp_unsigned_bytes: tx_req.rlp_unsigned().to_vec(),
+            rlp_bytes: tx_req.rlp_signed(&sig).to_vec(),
+            v: sig.v,
+            r: sig.r,
+            s: sig.s,
+            // other fields are irrelevant to get the sign_data()
+            ..Default::default()
+        };
 
-        SignData {
-            signature: (sig_r, sig_s, v),
-            pk,
-            msg: msg.into(),
-            msg_hash,
-        }
+        let sign_data = tx.sign_data().unwrap();
+        assert_eq!(sign_data.get_addr(), address!("0x7e5f4552091a69125d5dfcb7b8c2659029395bdf"));
+
+        sign_data
     };
 }
